Refactoring trie functions
